Abstract VROVisionEngine from body tracker
The VROVisionEngine handles the processing of
of ARKit frames — cropping and scaling them then
running them through an MLModel on a separate
thread — then returns them to its assigned delegate.

This commit pulls this functionality up from the
VROBodyTrackeriOS, so that it can (soon) be shared
with the body meshing algorithms.