Resample UV maps to render human meshes
Body mesher implementation that reads off a
test UV map (not direct ML output, yet), to
demonstrate rendering a resampled human body
mesh.

To accomplish this, we improted the cnpy library
to read Numpy arrays, and added the required
Numpy arrays from the ViroPoseNet library. These
arrays contain data like the texcoord to vertex
mapping, faces, and more — all of which are required
to correctly resample the mesh.