glTF metallic and roughness factor support
Metallic and roughness factors were not supported
by our glTF system. This adds support by:

1. Ensuring that if a roughness_metalness texture is
given, the roughness will be multiplied by the
roughness factor and the metalness will be multiplied
by the metalness factor. This is accomplished via
new intensity uniforms.

2. Ensuring that if no roughness_metalness texture
is given, the roughness and metalness factors will
apply to the entirety of the surface. This is accomplished
by setting the roughness and metalness uniform
colors.