VIRO-3701: immediately update umbrella bounds
The application thread was not immediately updating the
umbrella bounds when a node’s transforms were changed,
because this required traversing the entire scene graph.

This commit fixes this by updating these bounds on the
application thread whenever node properties are changed.
This means the umbrella bounding boxes for all objects in
the scene graph are always and immediately correct on
the application thread.

Note that the process used here could be made much more
efficiient (e.g. going from bottom up updating the bounds
instead of top-down, and/or batching updates together).

We use an inefficient top-down approach for now
because of ease of implementation (and AMNH needs this
fix quickly).