Refactor release test startup to prevent idle timeouts
Release tests timed out (failed to start) around 25% of
the time due to issues with the test framework waiting
for the Viro activity to idle before starting the test.

Fix this by making the test framework pause the
renderer after view construction, then resume after
the tests are ready to run.

Also remove the glInitialized check as it didn’t seem
to be necessary anymore (and was causing a different
deadlock).