Share vertex buffers across geometries
The GLTF spec contains the ability to share buffers
across different geometries, but Viro was not able
to do this, resulting in some GLTF models failing due
to out of memory errors on Viro.

This fix adds the ability to share a vertex buffer through
the new VROVertexBuffer class. This buffer is created
by the GLTF loader and cached, and its data is uploaded
only once to the GPU.

There currently exists a memory leak in that these VBOs
are not all being correctly discarded, which is addressed
in the next commit.