Immediately inject bounding box from Geo into Node
The UI thread did not have access to the latest
bounding box immediately after a Geometry was
injected into a Node. This caused issues when
users wanted to immediately get the bounds of a
Node on the UI thread.

To fix this, inject the bounding box of a Geometry
into its parent Node as soon as its set on the Node.
This also enables us to clean up the API and remove
the atomic bounding box that we previously used on
VROGeometry.