Compute local Node space bounding boxes
Developers occasionally may need bounding boxes
returned in local space instead of just world space.
This commit computes and stores local Node space
bounding boxes, accessible from both the renderer
thread and the application (UI) thread.

To faciliate this, the local transform is now also stored
in addition to the full world transform, for each Node.

We also store the “umbrella” local space bounding box,
which is the bounding box of a Node, and all of its
children, in the coordinate space of said Node. This
means all the children are transformed by their
respective transforms, up to but _not_ including the
tarnsform of the Node whose bounding box we are
computing.