Recompute bounding box after billboarding
This was actually already happening as a result of parent
nodes forcing child nodes to recompute bounding boxes
after constraints, but ensure it happens on leaf nodes as
well.

Note the previous concern with implementing this was that
the axis-aligned bounding boxes would result in hit areas
that are too large. We address this concern by dynamically
switching on high accuracy hit testing for billboarded objects.

This should be ok because billboarded objects are generally
low-poly (it’s one of the reasons billboards exist, to make
low-poly objects with textures substitute in for high
poly-objects!)