Allow the Animation of GLTF models with partial skeletal animations
GLTF skeletal animations need not necessairly have a complete
animation set that matches every bone in the skinner. For example
a humanoid skeleton may just have a wrist-bone animation.

This patch encounts for this situation and compensates for missing
animation bone data by using the original transforms of the 3D
model's node. By doing so, we are still able to re-construct
a complete VROSkeletalAnimation.