Simplified Animation Processing for GLTF models
- Simplified animation processing for GLTF models. We now convert each
gLTF animation channel into a VROKeyframe Animation to be played in
a parallel animation chain for a single gLTF animation. This helps
mitigate the issue where a single gLTF animation can have multiple
channels with different input time data.

- Removed any flattening of animation optimizations during the processing
of keyframe animations as it was not necessary since keyframes can be
played in parallel chains. This also helped removed bugs / issues and
reduced complexity of animations with multiple channels with
different input time data.

- Finally animation flattening is only applied for skeletal animations.
This is done to temporarily mitigate the need of layered skeletal
animations needed for skeletal animations with multi-channels with
different input time data. This mitigation is abstracted away into
a single flattenSkeletalKeyframeAnimations function that can be simply
removed once the skeletal layering feature is implemented for gLTF.

- Change is verified with all react-native gltf test models and as well
as render gltf test models.