Mitigate bug related to react navigation drawer
The react navigation drawer may be creating and
destroying the scene quickly, which can lead to a
case where we create a ViroViewARCore and then
dispose of it before the renderer’s onSurfaceCreated
is invoked. This can lead to a situation where the view
is not null but the view’s surface is null, causing the
crash.